requirements.txt
groq
python-dotenv
rich
streamlit
# transformers
# sentencepiece
# torch

